Working discrete Q learning implementation
After some unstable attempts with a NN-based policy, I switched the code
around to be able to learn a discrete Q table. This finally works as
expected: in a few iterations, the Q table fills up with reasonable
values and the length of episodes increases.

As reward, I use +1/0 depending on the angle. To get really good
results, using gamma>0 is needed - otherwise the policy will only learn
some real action around the state-space-cells that are adjecent to the
cells that get reward 0 - which might be too late to recover from the
soon-to-be-fully-dropped pole.

The idea of this rewrite was to be able to swap-out the NN policy with
the discrete policy. However, this doesn't fully work yet - the
simulation code currently still needs the discretization, and the
parameters for NN training need to be moved directly into the
deep_q_policy.py file. I will refactor stuff soon such that both are
compatible again.